# Update hold packages in yum.