### What does this PR do?

Similar to #30777, now using yum salt will update/upgraded held packages.

### What issues does this PR fix or reference?

#31566

### Previous Behavior

```bash
[root@salt ~]# yum versionlock list
Loaded plugins: fastestmirror, versionlock
0:epel-release-7-6.*
0:python-perf-3.10.0-327.18.2.el7.*
versionlock list done
```

```bash
[root@salt ~]# yum list updates
Loaded plugins: fastestmirror, versionlock
Loading mirror speeds from cached hostfile
 * base: mirrors.dcarsat.com.ar
 * centosplus: centos.zeotec.cl
 * epel: mirror.globo.com
 * epel-debuginfo: epel.gtdinternet.com
 * epel-source: mirror.globo.com
 * epel-testing: mirror.globo.com
 * epel-testing-debuginfo: epel.gtdinternet.com
 * epel-testing-source: mirror.globo.com
 * extras: mirror.globo.com
 * fasttrack: mirrors.dcarsat.com.ar
 * updates: mirror.globo.com
```

```yml
[root@salt ~]# cat /srv/salt/source.sls
epel-release:
  pkg.installed:
    - hold: True
    - version: 7-7

python-perf:
  pkg.installed:
    - hold: True
    - version: 3.10.0-327.18.2.el7.centos.plus
```

```yml
[root@salt ~]# salt '*' state.apply source
salt:
----------
          ID: epel-release
    Function: pkg.installed
      Result: False
     Comment: Error occurred installing package(s). Additional info follows:

              errors:
                  - Loaded plugins: fastestmirror, versionlock
                    Loading mirror speeds from cached hostfile
                     * base: mirrors.dcarsat.com.ar
                     * centosplus: centos.zeotec.cl
                     * epel: mirror.globo.com
                     * epel-debuginfo: epel.gtdinternet.com
                     * epel-source: mirror.globo.com
                     * epel-testing: mirror.globo.com
                     * epel-testing-debuginfo: epel.gtdinternet.com
                     * epel-testing-source: mirror.globo.com
                     * extras: mirror.globo.com
                     * fasttrack: mirrors.dcarsat.com.ar
                     * updates: mirror.globo.com
                    No package epel-release-7-7 available.
                    Error: Nothing to do
     Started: 07:17:13.723757
    Duration: 2127.424 ms
     Changes:
----------
          ID: python-perf
    Function: pkg.installed
      Result: False
     Comment: Error occurred installing package(s). Additional info follows:

              errors:
                  - Loaded plugins: fastestmirror, versionlock
                    Loading mirror speeds from cached hostfile
                     * base: mirrors.dcarsat.com.ar
                     * centosplus: centos.zeotec.cl
                     * epel: mirror.globo.com
                     * epel-debuginfo: epel.gtdinternet.com
                     * epel-source: mirror.globo.com
                     * epel-testing: mirror.globo.com
                     * epel-testing-debuginfo: epel.gtdinternet.com
                     * epel-testing-source: mirror.globo.com
                     * extras: mirror.globo.com
                     * fasttrack: mirrors.dcarsat.com.ar
                     * updates: mirror.globo.com
                    No package python-perf-3.10.0-327.18.2.el7.centos.plus available.
                    Error: Nothing to do
     Started: 07:17:15.851417
    Duration: 1640.429 ms
     Changes:

Summary for salt
------------
Succeeded: 0
Failed:    2
------------
Total states run:     2
ERROR: Minions returned with non-zero exit code
```

### New Behavior

```bash
[root@salt ~]# yum versionlock list
Loaded plugins: fastestmirror, versionlock
0:python-perf-3.10.0-327.13.1.el7.*
0:epel-release-7-5.*
versionlock list done
```

```bash
[root@salt ~]# yum list updates
Loaded plugins: fastestmirror, versionlock
Loading mirror speeds from cached hostfile
 * base: mirrors.dcarsat.com.ar
 * centosplus: centos.zeotec.cl
 * epel: mirror.globo.com
 * epel-debuginfo: epel.gtdinternet.com
 * epel-source: mirror.globo.com
 * epel-testing: mirror.globo.com
 * epel-testing-debuginfo: epel.gtdinternet.com
 * epel-testing-source: mirror.globo.com
 * extras: mirror.globo.com
 * fasttrack: mirrors.dcarsat.com.ar
 * updates: mirror.globo.com
```

```yml
[root@salt ~]# cat /srv/salt/source.sls
epel-release:
  pkg.installed:
    - hold: True
    - version: 7-7

python-perf:
  pkg.installed:
    - hold: True
    - version: 3.10.0-327.18.2.el7.centos.plus
```

```
[root@salt ~]# salt '*' state.apply source
salt:
----------
          ID: epel-release
    Function: pkg.installed
      Result: True
     Comment: The following packages were installed/updated: epel-release=7-7
              Package epel-release is already set to be held.
     Started: 07:10:04.251636
    Duration: 6363.53 ms
     Changes:
              ----------
              epel-release:
                  ----------
                  new:
                      7-7
                  old:
                      7-5
----------
          ID: python-perf
    Function: pkg.installed
      Result: True
     Comment: 1 targeted package was installed/updated.
              Package python-perf is already set to be held.
     Started: 07:10:10.636655
    Duration: 10925.828 ms
     Changes:
              ----------
              python-perf:
                  ----------
                  new:
                      3.10.0-327.18.2.el7.centos.plus
                  old:
                      3.10.0-327.13.1.el7

Summary for salt
------------
Succeeded: 2 (changed=2)
Failed:    0
------------
Total states run:     2
```

### Tests written?

No

@terminalmage sorry to bother, what probably you want to review this. Mostly because is not easy to handle lheld packages in yu, since in *targets* when updating a single package,  package name contains version number to install and that didn't work  with versionlock.

Thanks.