@abednarik 2016-06-06
I will check those tests errors when I get some time. Sorry for any inconvenience.