@abednarik 2016-06-06
I think this is related on how [_check_versionlock](https://github.com/abednarik/salt/blob/918e16fc00fb97113ecacc761bec74ac95a3475d/salt/modules/yumpkg.py#L174-L188) works. If this plugin is not installed, install will fail.
 I think I can add a try/except block to verify plugin existence and enable this code only if that is true.

Any feedback/recommendation is welcome.
Cheers.