@abednarik 2016-06-10
Hi @cachedout 

 This needs to be more work due to the way that salt checks if the right plugin for version lock is installed.
I want to hear some feedback or suggestions on how to handle this. Maybe a flag similar the ones in modules where you try to important some python libraries and set a constant.

Thanks.