@terminalmage 2016-06-16
I've noted a couple problems with this PR. Additionally, it seems that this PR doesn't address the case of a package which needs to be downgraded.