@terminalmage 2016-06-16
I will try to complete this PR, I will open a PR against @abednarik's fork.