@abednarik 2016-06-16
Thanks @terminalmage. I'm here to help.

Also the problem is that this code doesn't check if lock plugin is already installed, so maybe we need to do a try/except block to enable/disable this code?