@terminalmage 2016-06-16
@abednarik yeah, that is a good point. I will try to open a PR against your fork in the next few hours, and we can both test after that.