@abednarik 2016-06-16
No problem @terminalmage. I will have a look once I get some free time. Much appreciated your help.