@pirxthepilot 2016-06-20
Sorry, only got a chance to test this now.

Seems to works on single packages, but not multiples.

```yml
[root@minion2 salt]# cat php.sls 
{%- set version = '5.5.36-1.w7' -%}

php-packages:
  pkg.installed:
    - pkgs:
      - php55w-cli: {{ version }}
      - php55w-common: {{ version }}
    - hold: True
```

```
[root@minion2 salt]# rpm -qa|grep php55w
php55w-cli-5.5.35-1.w7.x86_64
php55w-common-5.5.35-1.w7.x86_64
```

```
[root@minion2 salt]# yum versionlock list
Loaded plugins: fastestmirror, versionlock
0:php55w-common-5.5.35-1.w7.*
0:php55w-cli-5.5.35-1.w7.*
versionlock list done
```

```yml
local:
----------
          ID: php-packages
    Function: pkg.installed
      Result: False
     Comment: Error occurred installing package(s). Additional info follows:
              
              errors:
                  - Loaded plugins: fastestmirror, versionlock
                    Loading mirror speeds from cached hostfile
                     * base: mirror.scalabledns.com
                     * epel: mirror.prgmr.com
                     * extras: mirror.hostduplex.com
                     * updates: mirrors.usc.edu
                     * webtatic: us-east.repo.webtatic.com
                    No package php55w-common-5.5.36-1.w7 available.
                    No package php55w-cli-5.5.36-1.w7 available.
                    Error: Nothing to do
     Started: 11:39:04.392320
    Duration: 5263.21 ms
     Changes:   

Summary for local
------------
Succeeded: 0
Failed:    1
------------
Total states run:     1
```

Version report:
```
Salt Version:
           Salt: 2016.3.0rc3-2167-g918e16f
 
Dependency Versions:
           cffi: Not Installed
       cherrypy: Not Installed
       dateutil: Not Installed
          gitdb: Not Installed
      gitpython: Not Installed
          ioflo: Not Installed
         Jinja2: 2.8
        libgit2: Not Installed
        libnacl: Not Installed
       M2Crypto: 0.24.0
           Mako: Not Installed
   msgpack-pure: Not Installed
 msgpack-python: 0.4.7
   mysql-python: Not Installed
      pycparser: Not Installed
       pycrypto: 2.6.1
         pygit2: Not Installed
         Python: 2.7.5 (default, Nov 20 2015, 02:00:19)
   python-gnupg: Not Installed
         PyYAML: 3.11
          PyZMQ: 15.2.0
           RAET: Not Installed
          smmap: Not Installed
        timelib: Not Installed
        Tornado: 4.3
            ZMQ: 4.0.5
 
System Versions:
           dist: centos 7.2.1511 Core
        machine: x86_64
        release: 3.10.0-327.18.2.el7.x86_64
         system: Linux
        version: CentOS Linux 7.2.1511 Core
```

Thanks!