@abednarik 2016-06-20
Thanks for testing this. Will try to complete this. Too busy down here lately, sorry.