@terminalmage 2016-06-21
@abednarik https://github.com/abednarik/salt/pull/2 has been opened, please review and test, and merge when you're satisfied. Once you've merged my PR to your fork, we can proceed with getting this PR merged.