@terminalmage 2016-06-21
@pirxthepilot the PR I opened against @abednarik's fork should address the problem you were seeing, I tested using the ``pkgs`` param. Keep an eye out once @abednarik has reviewed and merged the code I submitted to his fork, and you can test using the updated code.