@abednarik 2016-06-21
Hi

 I merged the code and is working for me. The only think I noticed in this error in the master log

```python
Process Maintenance-4:
Traceback (most recent call last):
  File "/usr/lib64/python2.7/multiprocessing/process.py", line 258, in _bootstrap
    self.run()
  File "/root/salt/salt/utils/process.py", line 640, in _run
    return self._original_run()
  File "/root/salt/salt/master.py", line 234, in run
    salt.daemons.masterapi.clean_old_jobs(self.opts)
  File "/root/salt/salt/daemons/masterapi.py", line 187, in clean_old_jobs
    mminion.returners[fstr]()
  File "/root/salt/salt/returners/local_cache.py", line 425, in clean_old_jobs
    shutil.rmtree(t_path)
  File "/usr/lib64/python2.7/shutil.py", line 239, in rmtree
    onerror(os.listdir, path, sys.exc_info())
  File "/usr/lib64/python2.7/shutil.py", line 237, in rmtree
    names = os.listdir(path)
OSError: [Errno 2] No such file or directory: '/var/cache/salt/master/jobs/29'
```

Which I don't see how this is related to this PR. @pirxthepilot can you give it a new try? Many thanks @terminalmage  for doing this and help me out.

Cheers.