@terminalmage 2016-06-21
@abednarik I can't reproduce this, but it seems similar to a bug we had in one of the 2016.3.0 release candidates? If you want to test the code from the PR, you can use https://help.github.com/articles/checking-out-pull-requests-locally/ to do that. The command you'd run to save the PR as a branch is:

```
git fetch origin pull/2/head:mybranch
```