@abednarik 2016-06-21
Thanks @terminalmage. That;s what I did and is working  good. Maybe that errors is related to my local development environment.