@terminalmage 2016-06-21
Test suite failed to run so I've closed/reopened to trigger a new test run.