@terminalmage 2016-06-21
@abednarik Can you remove the extra blank line before ``to_install`` is instantiated? It's line 1042 of ``yumpkg.py``. It's part of what I added in the code that I contributed, and the linter doesn't like it.