@abednarik 2016-06-21
@terminalmage done. Hope I did it right, otherwise let me know.

Cheers.