@terminalmage 2016-06-21
Weird, looks like you squashed that into the previous commit? It wasn't the commit where the extra line originated, but the result will be the same, so I guess it doesn't matter :smile: