@abednarik 2016-06-21
Yes, sorry. I always squash multiple commits when is the same change.  but didn't change your commit or at least was not my intention. Heavy day here, 2 migrations with differents clients :D