@terminalmage 2016-06-21
It should have been done in an interactive rebase so you could have moved the linter commit up one line, and squashed it into ce5727c. It got squashed to the wrong commit, it ended up being added to the commit where I had edited ``salt/states/pkg.py``.

This really is a minor nitpick though, I'm not worried about it. The end result is the same (and the linter is happy!).