@abednarik 2016-06-21
Cool, next time I will do it in a better moment instead of running to do this.