@terminalmage 2016-06-21
@cachedout test failures look unrelated, can you confirm?